Fixing and investigating

+++ src/model_code/stochastic_processes/policy_states_belief.py
@@ -12,7 +12,7 @@ def expected_SRA_probs_estimation(policy_state, choice, lagged_choice, options):
     # If fresh retired, you stay one more year in the same policy state
     fresh_retired = (choice == 0) & (lagged_choice != 0)
     n_policy_states = options["n_policy_states"]
-    no_policy_change = jnp.zeros(n_policy_states)
+    no_policy_change = jnp.zeros(n_policy_states, dtype=float)
     no_policy_change = no_policy_change.at[policy_state].set(1)
 
     # Aggregate the two transition vectors
